+++ src/lib/ast/get-assets.js
@@ -1,11 +1,11 @@
 import nunjucks from 'nunjucks';
 
-import {ImportWrapper} from '../../import-wrapper/ImportWrapper';
 import {StaticExtension} from '../../public/static-extension/StaticExtension';
 import {getFirstExistedPath} from '../../utils/get-first-existed-path';
 import {getPossiblePaths} from '../../utils/get-possible-paths';
 import {isUniqueAsset} from '../../utils/is-unique-asset';
 import {ERROR_MODULE_NOT_FOUND} from '../constants';
+import {ImportWrapper} from '../import-wrapper/ImportWrapper';
 
 import {getAddNodeValue} from './get-add-node-value';
 import {getNodesValues} from './get-nodes-values';
