+++ src/lib/import-wrapper/is-not-symbol-or-literal.js
