+++ src/lib/import-wrapper/is-symbol.js
