+++ src/lib/import-wrapper/normalize-trailing-slash.spec.js
