+++ src/lib/import-wrapper/optimize-import-literals.js
